Releasing version - 1.2.0 - Excluded players
Added functionality of excluding players, it was done to address case when due to talent distribution, player is not good for buffing/dispelling or other reasons.
Added a new argument to /buffduty command to exclude players.
Added Readme section and examples describing how to exclude players.
Added screenshots with examples to repository section `docs`.
Fixed title message when posting to channel to be universal and applicable not only to mages.
Updated tests to support excluded players.
Changed `mages` term to `players` term in tests.