Assignment order improvements.
Added order parameter to allow players to be assigned in a specific order.
Added assignments players to their groups first as a priority.
Sequential assignments instead of buffer number based(1,2,3 instead of 1,3,6).
Max groups are no longer 8, but based on number of players in raid.
Added CODEOWNERS file to properly assign PRs.